+++ src/main/groovy/com/cedarsoftware/util/JsonHttpProxy.groovy
@@ -54,7 +54,6 @@ import static org.apache.http.entity.ContentType.APPLICATION_JSON
 @PropertySource(value='classpath:application.properties')
 class JsonHttpProxy implements CallableBean
 {
-
     @Value("#{\${ncube.proxy.cookiesToInclude:'JSESSIONID'}.split(',')}")
     private List<String> cookiesToInclude
 
@@ -69,9 +68,9 @@ class JsonHttpProxy implements CallableBean
     private final int numConnections
     private static final Logger LOG = LoggerFactory.getLogger(JsonHttpProxy.class)
 
-    JsonHttpProxy(String scheme, String hostname, int port, String context, String username = null, String password = null, int numConnections = 6)
+    JsonHttpProxy(HttpHost httpHost, String context, String username = null, String password = null, int numConnections = 6)
     {
-        httpHost = new HttpHost(hostname, port, scheme)
+        this.httpHost = httpHost
         proxyHost = null
         this.context = context
         this.username = username